requirements.txt
require

google_auth_oauthlib
google-api-python-client

atlassian-python-api==3.14.0
slack_sdk==3.11.2